assignment1/requirements.txt
pygame-ce 
PySide6
hydra-core == 1.3.2